% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{case_path_nonsmooth}
\alias{case_path_nonsmooth}
\title{Case-weight adjusted solution path for L2 regularized nonsmooth problem (quantile regression and svm)}
\usage{
case_path_nonsmooth(
  a,
  B,
  c,
  lam,
  alpha_0,
  alpha_1,
  j,
  beta_0_w0,
  beta_w0,
  theta_w0
)
}
\arguments{
\item{a}{A \eqn{n \times 1} vector and a^T = (a_1, ..., a_n)}

\item{B}{A \eqn{n \times p} matrix and B^T = (b_1, ..., b_n)}

\item{c}{A \eqn{n \times 1} matrix and c^T = (b_1, ..., b_n)}

\item{lam}{Regularization parameter for L2 penalty}

\item{alpha_0}{A scalar in the definition of f(r)}

\item{alpha_1}{A scalar in the definition of f(r)}

\item{j}{Index of the observation that is attached a weight}

\item{beta_0_w0}{A scalar, which is the true value of beta_{0,w} when w = w_0 = 1}

\item{beta_w0}{A \eqn{p \times 1} vector, which is the true value of beta_{w} when w = w_0 = 1}

\item{theta_w0}{A \eqn{n \times 1} vector, which is the true value of the dual variable when w = w_0 = 1}
}
\value{
W_vec A list of breakout points

Beta_0 True values of beta_{0,w} at breakout points

Beta True values of beta_{w} at breakout points
}
\description{
Path-following algorithm to exactly solve 
	(beta_{0,w}, beta_{w}) = argmin_{beta_0, beta} \sum_{i \neq j} f(g_i(beta_0, beta)) + w*f(g_{j}(beta_0, beta)) + lambda / 2 * \|beta\|_2^2
for 0 <= w <= 1, where g_i(beta_0, beta) = a_i beta_0 + b_i^T beta + c_i and f(r) = alpha_0 max(r, 0) + alpha_1 max(-r, 0)
}
\details{
This function will be called by function CaseInfluence_nonsmooth to generate case influence graph for each case.
}
