% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CaseInfluence_logistic}
\alias{CaseInfluence_logistic}
\title{Compute global influence for each case under logistic regression}
\usage{
CaseInfluence_logistic(
  X,
  Y,
  theta_0,
  t_max,
  lam,
  epsilon,
  method,
  influence_measure
)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix}

\item{Y}{A \eqn{n \times 1} response vector}

\item{theta_0}{A \eqn{p \times 1} vector that corresponds to full-data solution}

\item{t_max}{Range of solution path \eqn{[0, t_{\max}]}}

\item{lam}{Regularization parameter for L2 penalty}

\item{epsilon}{Pre-specified error tolerance of the entire path}

\item{method}{The method to use at each grid point, two options are "Newton" and "GD"}

\item{influence_measure}{The influence measure used to compute global influence for each case, two options are "FMD" and "BDLD"}
}
\value{
global_influence Global influence for each case

cook_distance Cook's distance for each case
}
\description{
Compute the global influence/Cook's distance for each case under logistic regression.
}
\details{
This function will be called by the main function Compute_CaseInflu_GLM, with class = "logistc".
}
