% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CaseInfluence_poisson}
\alias{CaseInfluence_poisson}
\title{Compute global influence for each case under Poisson regression}
\usage{
CaseInfluence_poisson(X, Y, theta_0, t_max, lam, epsilon, method)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix}

\item{Y}{A \eqn{n \times 1} response vector}

\item{theta_0}{A \eqn{p \times 1} vector that corresponds to full-data solution}

\item{t_max}{Range of solution path \eqn{[0, t_{\max}]}}

\item{lam}{Regularization parameter for L2 penalty}

\item{epsilon}{Pre-specified error tolerance of the entire path}

\item{method}{The method to use at each grid point, two options are "Newton" and "GD"}
}
\value{
global_influence Global influence for each case

cook_distance Cook's distance for each case
}
\description{
Compute the global influence/Cook's distance for each case under Poisson regression.
}
\details{
This function will be called by the main function Compute_CaseInflu_GLM, with class = "poisson".
}
