% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Compute_CaseInflu_GLM}
\alias{Compute_CaseInflu_GLM}
\title{Case influence assessment for L2 penalized GLM}
\usage{
Compute_CaseInflu_GLM(
  X,
  Y,
  lam,
  epsilon,
  t_max = 10,
  class = "logistic",
  method = "Newton",
  influence_measure = "FMD"
)
}
\arguments{
\item{X}{A n by p feature matrix}

\item{Y}{A n by 1 response vector}

\item{lam}{Regularization parameter for L2 penalty}

\item{epsilon}{Pre-specified error tolerance of the entire path}

\item{t_max}{Range of solution path [0, t_max]}

\item{class}{Specify the problem class with default value "logistic". Use "logistic" for logistic regression, and "poisson" for poisson regression}

\item{method}{Specify the method for computing the solution path. Default is "Newton". Use "GD" for gradient descent}

\item{influence_measure}{Specify the influence measure for computing global influence. For classification problem, "FMD" represents
functional margin difference and "BDLD" represents Binomial deviance loss diference. The default is "FMD". For regression problem, no
need to specify the influence_measure}
}
\value{
global_influence Global influence for each case

cook_distance Cook's distance for each case
}
\description{
Compute the global influence, which is based on case influence graph, for each case under L2 regularized logistic
regression. To compute the case influence graph for each case, we need to solve a solution path problem. In our algorithm,
the grid points are adaptively selected to save computation while maintaining the accuracy of the entire path. At each grid point,
the optimization problem can be solved by either Newton method or gradient descent method.
}
