% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Compute_LocalInflu_GLM}
\alias{Compute_LocalInflu_GLM}
\title{Local influence assessment for L2 penalized GLM}
\usage{
Compute_LocalInflu_GLM(
  X,
  Y,
  lam,
  class = "logistic",
  influence_measure = "FMD"
)
}
\arguments{
\item{X}{A n by p feature matrix}

\item{Y}{A n by 1 response vector}

\item{lam}{Regularization parameter for L2 penalty}

\item{class}{Specify the problem class with default value "quantile". Use "quantile" for quantile regression, and "svm" for support vector machine}

\item{influence_measure}{Specify the influence measure for computing local influence. For classification problem, "FMD" represents
functional margin difference and "BDLD" represents Binomial deviance loss difference. The default is "FMD". For regression problem, no
need to specify the influence_measure}
}
\value{
LocalInfluence_vec Local influence for each case
}
\description{
Compute the local influence for each case under L2 regularized GLM.
}
