% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LocalInfluence_logistic}
\alias{LocalInfluence_logistic}
\title{Compute local influence for each case under logistic regression}
\usage{
LocalInfluence_logistic(X, Y, theta_0, lam, influence_measure)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix}

\item{Y}{A \eqn{n \times 1} response vector}

\item{theta_0}{A \eqn{p \times 1} vector that corresponds to full-data solution}

\item{lam}{Regularization parameter for L2 penalty}

\item{influence_measure}{The influence measure used to compute global influence for each case, two options are "FMD" and "BDLD"}
}
\value{
LocalInfluence_vec Local influence for each case
}
\description{
Compute the local influence for each case under logistic regression.
}
\details{
This function will be called by the main function Compute_LocalInflu_GLM, with class = "logistic".
}
