% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LocalInfluence_nonsmooth}
\alias{LocalInfluence_nonsmooth}
\title{Compute local influence for each case under L2 regularized nonsmooth problem}
\usage{
LocalInfluence_nonsmooth(
  a,
  B,
  c,
  lam,
  alpha_0,
  alpha_1,
  beta_0_w0,
  beta_w0,
  theta_w0,
  model_class,
  influence_measure
)
}
\arguments{
\item{a}{A \eqn{n \times 1} vector and a^T = (a_1, ..., a_n)}

\item{B}{A \eqn{n \times p} matrix and B^T = (b_1, ..., b_n)}

\item{c}{A \eqn{n \times 1} matrix and c^T = (b_1, ..., b_n)}

\item{lam}{Regularization parameter for L2 penalty}

\item{alpha_0}{A scalar in the definition of f(r)}

\item{alpha_1}{A scalar in the definition of f(r)}

\item{beta_0_w0}{A scalar, which is the true value of beta_{0,w} when w = w_0 = 1}

\item{beta_w0}{A \eqn{p \times 1} vector, which is the true value of beta_{w} when w = w_0 = 1}

\item{theta_w0}{A \eqn{n \times 1} vector, which is the true value of the dual variable when w = w_0 = 1}

\item{model_class}{The model we consider, two options are "quantile" and "svm"}

\item{influence_measure}{The influence measure used to compute global influence for each case, two options are "FMD" and "BDLD". It is only used when
model_class = "svm"}
}
\value{
LocalInfluence_vec Local influence for each case
}
\description{
Compute the local influence for each case under L2 regularized nonsmooth problem (quantile regression and svm), 
and the procedure only depends on the full-data solution.
}
\details{
This function will be called by the main function Compute_LocalInflu_nonsmooth to compute local influence for each case.
}
