% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LocalInfluence_poisson}
\alias{LocalInfluence_poisson}
\title{Compute local influence for each case under Poisson regression}
\usage{
LocalInfluence_poisson(X, Y, theta_0, lam)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix}

\item{Y}{A \eqn{n \times 1} response vector}

\item{theta_0}{A \eqn{p \times 1} vector that corresponds to full-data solution}

\item{lam}{Regularization parameter for L2 penalty}
}
\value{
LocalInfluence_vec Local influence for each case
}
\description{
Compute the local influence for each case under Poisson regression.
}
\details{
This function will be called by the main function Compute_LocalInflu_GLM, with class = "poisson".
}
