% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{LOO_GLM}
\alias{LOO_GLM}
\title{LOO solution computation for L2 penalized GLM}
\usage{
LOO_GLM(
  X,
  Y,
  lam,
  obs_index,
  epsilon,
  t_max = 10,
  class = "logistic",
  method = "Newton"
)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix}

\item{Y}{A \eqn{n \times 1} response vector}

\item{lam}{Regularization parameter for L2 penalty}

\item{obs_index}{Index of the observation that is left out}

\item{epsilon}{Pre-specified error tolerance of the entire path}

\item{t_max}{Range of solution path \eqn{[0, t_{\max}]}}

\item{class}{Specify the problem class with default value "logistic". Use "logistic" for logistic regression, and "poisson" for poisson regression}

\item{method}{Specify the method for computing the solution path. Default is "Newton". Use "GD" for gradient descent}
}
\value{
The computed LOO solution
}
\description{
Compute the LOO solution based on the one-step Newton method, using the solution at the last grid point as the initialization
}
