% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nonsmooth_path}
\alias{nonsmooth_path}
\title{Path-following algorithm for L2 penalized nonsmooth problems}
\usage{
nonsmooth_path(X, Y, lam, obs_index, class = "quantile", tau = 0.5)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix}

\item{Y}{A \eqn{n \times 1} response vector}

\item{lam}{Regularization parameter for L2 penalty}

\item{obs_index}{Index of the observation that is attached a weight, ranging from 1 to n}

\item{class}{Specify the problem class with default value "quantile". Use "quantile" for quantile regression, and "svm" for support vector machine}

\item{tau}{The parameter for quantile regression, ranging between 0 and 1. No need to specify the value of it if choose class = "svm"}
}
\value{
W_vec A list of breakout points

Beta_0 True values of beta_{0,w} at breakout points

Beta True values of beta_{w} at breakout points
}
\description{
Compute exact solution path for case-weight adjusted problem 
\loadmathjax
\mjseqn{(\beta_{0,w}, \beta_{w}) = argmin_{\beta_0, \beta} \sum_{i \neq j} f(g_i(\beta_0, \beta)) + w*f(g_{j}(\beta_0, \beta)) + \lambda / 2 * \|\beta\|_2^2}
for \mjseqn{0 <= w <= 1}, where \mjseqn{g_i(\beta_0, \beta) = a_i \beta_0 + b_i^T \beta + c_i} and \mjseqn{f(r) = \alpha_0 \max(r, 0) + \alpha_1 \max(-r, 0).}
The exact solution path is shown to be piece-wise linear 
and the function will return all the breakpoints and their solutions.
}
