% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poisson_compute_LOO}
\alias{poisson_compute_LOO}
\title{Compute the LOO solution for L2 regularized poisson regression}
\usage{
poisson_compute_LOO(X, Y, theta_init, lam, case_index)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix}

\item{Y}{A \eqn{n \times 1} response vector}

\item{theta_init}{A \eqn{p \times 1} vector, which is the initial point}

\item{lam}{Regularization parameter for L2 penalty}

\item{case_index}{Index of the observation that is left out}
}
\value{
theta_loo Computed LOO solutions
}
\description{
Given the initial point, compute the LOO solution for L2 regularized poisson regression by one-step Newton method
}
