% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{poisson_Newton_low_dimension}
\alias{poisson_Newton_low_dimension}
\title{Case-weight adjusted solution path for poisson regression Newton method (\eqn{p \le n})}
\usage{
poisson_Newton_low_dimension(X, Y, theta_0, t_max, lam, case_index, epsilon, L)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix}

\item{Y}{A \eqn{n \times 1} response vector}

\item{theta_0}{A \eqn{p \times 1} vector that corresponds to full-data solution}

\item{t_max}{Range of solution path \eqn{[0, t_{\max}]}}

\item{lam}{Regularization parameter for L2 penalty}

\item{case_index}{Index of the observation that is attached a weight}

\item{epsilon}{Pre-specified error tolerance of the entire path}

\item{L}{Lipschitz constant for the gradient of the loss function}
}
\value{
t_vec Adaptively selected grid points

theta Generated soultion path at grid points

alpha_t Step size between grid points
}
\description{
Generate solution path for case-weight adjusted problem by adopting Newton method at each selected grid points,
which are adaptively selected to save computation while maintaining the accuracy of the entire path.
}
\details{
This function will be called by the function CaseInfluence_poisson to generate case influence graph for each case, with method = "Newton".
}
