% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{smooth_path}
\alias{smooth_path}
\title{Path-following algorithm for L2 penalized Generalized linear models}
\usage{
smooth_path(
  X,
  Y,
  lam,
  obs_index,
  epsilon,
  method = "Newton",
  t_max = 10,
  class = "logistic"
)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix}

\item{Y}{A \eqn{n \times 1} response vector}

\item{lam}{Regularization parameter for L2 penalty}

\item{obs_index}{Index of the observation that is attached a weight, ranging from 1 to n}

\item{epsilon}{Pre-specified error tolerance of the entire path}

\item{method}{Specify the method for computing the solution path. Default is "Newton". Use "GD" for gradient descent}

\item{t_max}{Range of solution path \eqn{[0, t_{\max}]}}

\item{class}{Specify the problem class with default value "logistic". Use "logistic" for logistic regression, and "poisson" for poisson regression}
}
\value{
t_vec Adaptively selected grid points

theta Generated solution path at grid points

alpha_t Step size between grid points
}
\description{
Compute the global influence, which is based on case influence graph, for each case under \eqn{\ell_2} regularized Generalized linear
Models. To compute the case influence graph for each case, we need to solve the following solution path problem: 
\loadmathjax
\mjdeqn{\theta(t) = argmin_{\theta}  e^{-t} L_{j}(\theta) + \sum_{i \neq j} L_{i}(\theta) + \lambda / 2 * \|\theta\|_2^2 \text{ for } t \in [0, t_{max}]}{ASCII representation}
In our algorithm, the grid points are adaptively selected to save computation while maintaining the accuracy of the entire path.
At each grid point, the optimization problem can be solved by either Newton method or gradient descent method.
}
